+++ SessionManager/web/admin/sharedfolders.php
@@ -21,8 +21,16 @@
 require_once(dirname(__FILE__).'/includes/core.inc.php');
 require_once(dirname(__FILE__).'/includes/page_template.php');
 
+if (! checkAutorization('viewSharedFolders'))
+	redirect('index.php');
+
+
 if (isset($_REQUEST['action'])) {
 	if ($_REQUEST['action'] == 'rename' && isset($_REQUEST['id'])) {
+		if (! checkAutorization('manageSharedFolders'))
+			redirect();
+
+
 		if (isset($_REQUEST['sharedfolder_name'])) {
 			$sharedfolder = Abstract_SharedFolder::load($_REQUEST['id']);
 			if (is_object($sharedfolder)) {
@@ -45,6 +53,8 @@ if (isset($_REQUEST['action'])) {
 function show_default() {
 	$sharedfolders = SharedFolders::getAll();
 
+	$can_manage_sharedfolders = isAutorized('manageSharedFolders');
+
 	page_header();
 
 	echo '<div id="sharedfolders_div">';
@@ -56,21 +66,25 @@ function show_default() {
 	foreach ($sharedfolders as $sharedfolder) {
 		echo '<tr>';
 		echo '<td><a href="sharedfolders.php?action=manage&amp;id='.$sharedfolder->id.'">'.$sharedfolder->name.'</a></td>';
-		echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder?').'\');">';
-		echo '<input type="hidden" name="name" value="SharedFolder" />';
-		echo '<input type="hidden" name="action" value="del" />';
-		echo '<input type="hidden" name="id" value="'.$sharedfolder->id.'" />';
-		echo '<input type="submit" value="'._('Delete this shared folder').'" />';
-		echo '</form></td>';
+		if ($can_manage_sharedfolders) {
+			echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder?').'\');">';
+			echo '<input type="hidden" name="name" value="SharedFolder" />';
+			echo '<input type="hidden" name="action" value="del" />';
+			echo '<input type="hidden" name="id" value="'.$sharedfolder->id.'" />';
+			echo '<input type="submit" value="'._('Delete this shared folder').'" />';
+			echo '</form></td>';
+		}
 		echo '</tr>';
 	}
 
-	echo '<tr><form action="actions.php" method="post"><td>';
-	echo '<input type="hidden" name="name" value="SharedFolder" />';
-	echo '<input type="hidden" name="action" value="add" />';
-	echo '<input type="text" name="sharedfolder_name" value="" />';
-	echo '</td><td><input type="submit" value="'._('Create this shared folder').'" /></td>';
-	echo '</form></tr>';
+	if ($can_manage_sharedfolders) {
+		echo '<tr><form action="actions.php" method="post"><td>';
+		echo '<input type="hidden" name="name" value="SharedFolder" />';
+		echo '<input type="hidden" name="action" value="add" />';
+		echo '<input type="text" name="sharedfolder_name" value="" />';
+		echo '</td><td><input type="submit" value="'._('Create this shared folder').'" /></td>';
+		echo '</form></tr>';
+	}
 
 	echo '</table>';
 	echo '</div>';
@@ -100,6 +114,8 @@ function show_manage($sharedfolder_id_) {
 			$available_groups[] = $group;
 	}
 
+	$can_manage_sharedfolders = isAutorized('manageSharedFolders');
+
 	page_header();
 
 	echo '<div id="sharedfolders_div">';
@@ -113,12 +129,16 @@ function show_manage($sharedfolder_id_) {
 	echo '<tr><td>';
 	echo _('Name').': ';
 	echo '</td><td>';
-	echo '<form action="sharedfolders.php" method="post">';
-	echo '<input type="hidden" name="action" value="rename" />';
-	echo '<input type="hidden" name="id" value="'.$sharedfolder->id.'" />';
+	if ($can_manage_sharedfolders) {
+		echo '<form action="sharedfolders.php" method="post">';
+		echo '<input type="hidden" name="action" value="rename" />';
+		echo '<input type="hidden" name="id" value="'.$sharedfolder->id.'" />';
+	}
 	echo '<input type="text" name="sharedfolder_name" value="'.$sharedfolder->name.'" />';
-	echo ' <input type="submit" value="'._('Rename').'" />';
-	echo '</form>';
+	if ($can_manage_sharedfolders) {
+		echo ' <input type="submit" value="'._('Rename').'" />';
+		echo '</form>';
+	}
 	echo '</td></tr>';
 
 	echo '</table>';
@@ -133,17 +153,19 @@ function show_manage($sharedfolder_id_) {
 	foreach ($used_groups as $group) {
 		echo '<tr>';
 		echo '<td><a href="usersgroup.php?action=manage&amp;id='.$group->getUniqueID().'">'.$group->name.'</a></td>';
-		echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder access?').'\');">';
-		echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
-		echo '<input type="hidden" name="action" value="del" />';
-		echo '<input type="hidden" name="sharedfolder_id" value="'.$sharedfolder->id.'" />';
-		echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
-		echo '<input type="submit" value="'._('Delete access to this shared folder').'" />';
-		echo '</form></td>';
+		if ($can_manage_sharedfolders) {
+			echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder access?').'\');">';
+			echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
+			echo '<input type="hidden" name="action" value="del" />';
+			echo '<input type="hidden" name="sharedfolder_id" value="'.$sharedfolder->id.'" />';
+			echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
+			echo '<input type="submit" value="'._('Delete access to this shared folder').'" />';
+			echo '</form></td>';
+		}
 		echo '</tr>';
 	}
 
-	if (count($available_groups) > 0) {
+	if (count($available_groups) > 0 and $can_manage_sharedfolders) {
 		echo '<tr><form action="actions.php" method="post"><td>';
 		echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
 		echo '<input type="hidden" name="action" value="add" />';
